requirements.txt
slack-bolt
python-dotenv
snowflake
snowflake-snowpark-python
requests
pandas
numpy
matplotlib